4.0.0.38

+++ MfcTools/LoadModifierPage.cpp
@@ -43,7 +43,7 @@ static DWORD CALLBACK StreamInCtrl(DWORD_PTR dwCookie, LPBYTE pbBuff, LONG cb, L
 
    for ( int i = 0; i < cb; i++ )
    {
-      *(pbBuff+i) = psBuffer->GetAt(i);
+      *(pbBuff+i) = (BYTE)psBuffer->GetAt(i);
    }
    *pcb = cb;
    *psBuffer = psBuffer->Mid(cb);
